# backend/src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/quiz-sessions/{sessionId}": {
      "get": {
        "operationId": "QuizSessionController_getQuizSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetQuizSessionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Quiz-sessions"
        ]
      },
      "patch": {
        "operationId": "QuizSessionController_updateQuizSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetQuizSessionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Quiz-sessions"
        ]
      }
    },
    "/quiz-sessions/create": {
      "post": {
        "operationId": "QuizSessionController_createQuizSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetQuizSessionLiteResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Quiz-sessions"
        ]
      }
    },
    "/quiz-sessions/response": {
      "post": {
        "operationId": "QuizSessionController_createQuizResponse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizResponseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetQuizResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Quiz-sessions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Interview Ready",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "GetQuizSessionLiteResponse": {
        "type": "object",
        "properties": {
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "ABANDONED"
            ],
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ]
          }
        },
        "required": [
          "completedAt",
          "id",
          "title",
          "status",
          "startedAt",
          "user"
        ]
      },
      "GetQuizResponseItemQuestionCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "GetQuizResponseItemQuestionAnswerOption": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCorrect",
          "text"
        ]
      },
      "GetQuizResponseItemQuestion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "explanation": {
            "type": "object"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "MULTIPLE_CHOICE",
              "FREE_TEXT"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR"
            ]
          },
          "aiGenerated": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/GetQuizResponseItemQuestionCategory"
          },
          "answerOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetQuizResponseItemQuestionAnswerOption"
            }
          }
        },
        "required": [
          "id",
          "text",
          "explanation",
          "type",
          "difficulty",
          "aiGenerated",
          "createdAt",
          "updatedAt",
          "category",
          "answerOptions"
        ]
      },
      "GetQuizResponseItem": {
        "type": "object",
        "properties": {
          "textAnswer": {
            "type": "string",
            "nullable": true
          },
          "feedback": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean",
            "nullable": true
          },
          "score": {
            "type": "number",
            "nullable": true
          },
          "answeredAt": {
            "format": "date-time",
            "type": "string"
          },
          "question": {
            "$ref": "#/components/schemas/GetQuizResponseItemQuestion"
          },
          "answersIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "textAnswer",
          "feedback",
          "id",
          "isCorrect",
          "score",
          "answeredAt",
          "question",
          "answersIds"
        ]
      },
      "GetQuizSessionResponse": {
        "type": "object",
        "properties": {
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "responses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetQuizResponseItem"
            }
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "ABANDONED"
            ],
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ]
          }
        },
        "required": [
          "completedAt",
          "responses",
          "id",
          "title",
          "status",
          "startedAt",
          "user"
        ]
      },
      "CreateQuizSessionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateQuizSessionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "ABANDONED"
            ]
          }
        },
        "required": [
          "title",
          "status"
        ]
      },
      "CreateQuizResponseAnswerRequest": {
        "type": "object",
        "properties": {
          "answerOptionId": {
            "type": "string",
            "pattern": "^[a-z0-9]{24,}$"
          }
        },
        "required": [
          "answerOptionId"
        ]
      },
      "CreateQuizResponseRequest": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateQuizResponseAnswerRequest"
            }
          },
          "sessionId": {
            "type": "string",
            "pattern": "^[a-z0-9]{24,}$"
          },
          "questionId": {
            "type": "string",
            "pattern": "^[a-z0-9]{24,}$"
          },
          "textAnswer": {
            "type": "string",
            "nullable": true
          },
          "feedback": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "answers",
          "sessionId",
          "questionId",
          "textAnswer",
          "feedback"
        ]
      },
      "GetQuizResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "object"
          },
          "score": {
            "type": "object"
          },
          "answeredAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCorrect",
          "score",
          "answeredAt"
        ]
      },
...
    }
  }
}